refactor(ui): extract QuickActions component and standardize template & prompt action buttons
